Add activation steering and SIRA analysis pipeline for LIBERO tasks
- Steering evaluation (task-specific and general model) with configurable
  norm preservation, per-layer coefficient, and Spearman correlation analysis
- SIRA correlation analysis and vector similarity tools
- Full 3-phase pipeline: collect hidden states, compute vectors, steered inference
- SLURM submission and setup scripts for H100 cluster
- Pin diffusers version and save ground truth actions

+++ prismatic/training/train_utils.py
@@ -7,7 +7,7 @@ from prismatic.vla.constants import ACTION_DIM, ACTION_TOKEN_BEGIN_IDX, IGNORE_I
 
 def get_current_action_mask(token_ids):
     # Create a tensor marking positions of IGNORE_INDEX
-    newline_positions = token_ids != IGNORE_INDEX
+    newline_positions = (token_ids != IGNORE_INDEX).to(torch.int64)
 
     # Calculate cumulative sum to identify regions between newlines
     cumsum = torch.cumsum(newline_positions, dim=1)
@@ -24,7 +24,7 @@ def get_current_action_mask(token_ids):
 
 def get_next_actions_mask(token_ids):
     # Create a tensor marking positions of IGNORE_INDEX
-    newline_positions = token_ids != IGNORE_INDEX
+    newline_positions = (token_ids != IGNORE_INDEX).to(torch.int64)
 
     # Calculate cumulative sum to identify regions between newlines
     cumsum = torch.cumsum(newline_positions, dim=1)
